[HOTSPOT-74] refactor: 구조 변경 및 리뷰 반영
