+++ apps/web/src/entities/report/index.ts
@@ -1,4 +1,3 @@
 export { useUsageReport } from './api/useUsageReport';
 export type { ReportUser, UsageResponse } from './model/type';
 export { ServiceReport } from './ui/ServiceReport';
-export { UserChip } from './ui/UserChip';
