+++ apps/web/src/entities/user/index.ts
@@ -1,0 +1,1 @@
+export { UserSelector } from './ui/UserSelector';
