+++ apps/web/src/shared/ui/index.ts
@@ -1,3 +1,4 @@
 export { BottomSheet } from './bottom-sheet/BottomSheet';
 export { Chip } from './chip/Chip';
-export { MonthPicker } from './month-picker/MonthPicker';
+export { DateNavigation } from './date/DateNavigation';
+export { MonthDaySelector } from './date/MonthDaySelector';
