+++ packages/ui/src/lib/getRoundedMax.ts
